headless mpc: start pulling QT out of MultiPointController and Worker
