+++ software/control/widgets.py
@@ -4,7 +4,8 @@ import sys
 from typing import Optional
 
 import squid.logging
-from control.core.core import TrackingController, MultiPointController, LiveController
+from control.core.core import TrackingController, LiveController
+from control.core.multi_point_controller import MultiPointController
 from control.microcontroller import Microcontroller
 from control.piezo import PiezoStage
 import control.utils as utils
